test, remove duplicated tests inbetween AFD module test and AFT tests.
